pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "newsticker"
version = "0.0.1"
authors = [
  { name="Christian Karrié", email="ckarrie@gmail.com" },
]
description = "Newsticker"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'django',
  'django-treebeard'
]

[project.urls]
Homepage = "https://github.com/ckarrie/ckw-newsticker"
Issues = "https://github.com/ckarrie/ckw-newsticker/issues"
